refactor(domain): 중복 코드 제거를 위해 상속 적용
Dealer와 Player 클래스에서 중복으로 발생하는 코드를 부모 클래스 Participant를 생성 후, 옮겨서 중복을 제거했습니다.
점수 계산 로직에서 A가 존재하지 않아도 calculateAScore()메소드가 호출되는 문제를 해결했습니다.
이제 카드 점수 계산 테스트를 Player에서 Participant로 수정했습니다.
