+++ src/test/java/domain/ParticipantTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class PlayerTest {
+public class ParticipantTest {
 
     private static Stream<Arguments> safeScoreCards() {
         return Stream.of(
@@ -27,9 +27,9 @@ public class PlayerTest {
     @ParameterizedTest
     @MethodSource("safeScoreCards")
     public void 카드_점수_합계를_정상적으로_계산한다(List<Card> cards, int expectedValue) {
-        Player player = new Player("스타크");
-        cards.forEach(player::addCard);
-        int score = player.calculateScore();
+        Participant participant = new Participant();
+        cards.forEach(participant::addCard);
+        int score = participant.calculateScore();
 
         assertThat(score).isEqualTo(expectedValue);
     }
@@ -37,10 +37,10 @@ public class PlayerTest {
     @DisplayName("카드 점수 합계가 21을 초과하면 버스트를 판정한다.")
     @Test
     public void 카드_점수_합계가_21을_넘으면_버스트를_판정한다() {
-        Player player = new Player("스타크");
+        Participant participant = new Participant();
         List<Card> burstCards = List.of(new Card(Rank.JACK, Suit.CLOVER), new Card(Rank.QUEEN, Suit.CLOVER), new Card(Rank.TWO, Suit.CLOVER));
-        burstCards.forEach(player::addCard);
+        burstCards.forEach(participant::addCard);
 
-        assertThat(player.isBurst()).isTrue();
+        assertThat(participant.isBurst()).isTrue();
     }
 }
